FIX wrong success for an erronous fetch, after an abort of a legit fetch
This bug can be manifested when a pending fetch of a legit resource (200 response)
is aborted and afterwards a new fetch that leads to error (e.g. 404) is issued.
At the time, the second request finishes the lastValue has the aborted data of
the previous fetch (callback is executed even on aborted), YIELDING the second (normally
erroneus response) to be signaled as successfull.